+++ lib/activerecord-postgis-array.rb
@@ -1,0 +1,3 @@
+require 'activerecord-postgis-array/version'
+require 'activerecord-postgis-array/active_record'
+require 'activerecord-postgis-array/arel'