+++ lib/activerecord-postgis-array/active_record.rb
@@ -1,0 +1,4 @@
+require 'activerecord-postgis-array/active_record/sanitization'
+require 'activerecord-postgis-array/active_record/connection_adapters'
+require 'activerecord-postgis-array/active_record/schema_dumper'
+require 'activerecord-postgis-array/active_record/relation'