+++ lib/activerecord-postgis-array/active_record/connection_adapters.rb
@@ -1,0 +1,1 @@
+require 'activerecord-postgis-array/active_record/connection_adapters/postgres_adapter'