+++ lib/activerecord-postgis-array/active_record/relation.rb
@@ -1,0 +1,2 @@
+require 'activerecord-postgis-array/active_record/relation/query_methods'
+require 'activerecord-postgis-array/active_record/relation/predicate_builder'