+++ lib/activerecord-postgis-array/arel.rb
@@ -1,0 +1,3 @@
+require 'activerecord-postgis-array/arel/nodes'
+require 'activerecord-postgis-array/arel/predications'
+require 'activerecord-postgis-array/arel/visitors'