+++ lib/activerecord-postgis-array/arel/nodes.rb
@@ -1,0 +1,2 @@
+require 'activerecord-postgis-array/arel/nodes/array_nodes'
+require 'activerecord-postgis-array/arel/nodes/contained_within'