+++ lib/activerecord-postgis-array/arel/visitors.rb
@@ -1,0 +1,2 @@
+require 'activerecord-postgis-array/arel/visitors/visitor'
+require 'activerecord-postgis-array/arel/visitors/to_sql'