+++ lib/postgres_ext/active_record/relation.rb
@@ -1,2 +1,0 @@
-require 'postgres_ext/active_record/relation/query_methods'
-require 'postgres_ext/active_record/relation/predicate_builder'