+++ lib/postgres_ext/arel.rb
@@ -1,3 +1,0 @@
-require 'postgres_ext/arel/nodes'
-require 'postgres_ext/arel/predications'
-require 'postgres_ext/arel/visitors'